Update BuzzNumber.java
Hello, This is my first commit... Im sorry if i did it wrong. I added functionality so that if number entered is negative it will prompt user to re-enter a positive number. Also added funtionality in a try catch block so that if user enters a non number it prompts user to enter a number before it would have errored out.