Create majorityElementHashhmap.java
 HashMap has ability to efficiently store and manage data by associating keys with values and it also improves the Time Complexity. The question which is solved in the provided java code is a very known question and it builds a strong foundation for the topic hashmap. Question follows:
 Given an Array of size n, find all the elements that appear more than n/3 times Array={1, 3, 2, 5, 1, 3, 1, 5, 1 }.
Using two nested for loops to solve this problem would result in a less efficient solution with a time complexity of O(n^2), making it slower for large input arrays compared to the HashMap-based solution, which has a time complexity of O(n). This code uses a HashMap to efficiently count how many times each unique element appears in an array. It then checks if any of these elements appear more than one-third of the total elements in the array and prints them if they do. This demonstrates how HashMaps are valuable for tracking and analyzing data efficiently in real-world programming scenarios.